Remove the namespaces.k8s and rename namespaces.nsm
This switches to use the built in .Release.Namespace variable from
helm so that it will use whatever namespace the deployment is in. The
main benefit here is that there isn't any confusion about how to deploy
this to a specific namespace, its unified to be the one helm is told to
use.

We also remove references to "nsm" and just call it the pure namespace
for the storage layers namespacing.