Add CSI plugin directory (#102)
* Add CSI plugin directory

* Now we have two helm-charts and two Operators for the FlexVolume and CSI plugin respectively.
Update the README of the Operators.

* Updates using PSO Operator is broken as of 2.5.0+
This is because the chart now creates a ClusteRole "pure-provisioner-clusterrole" and we cannot Get the ClusterRole resource by name before it is created.
Earlier we used the system:persistent-volume-provisioner ClusterRole which already exists. Hence it was okay to add "get" privilege to the Clusterrole by resource name.
Because of this issue updates fail i.e. Operator does not apply changes in values.yaml.
The fix is to add "get" privilege to ClusterRole without using the resource-name qualifier.

* Remove CRD creation from the chart
* use correct version tag